Simplify lib/parser structure
Instead of wrapping lib/parser/juttle-parser.js into a nearly empty
lib/parser/index.js module, let's put all the functionality into
lib/parser/index.js directly.
