+++ lib/compiler/autocompleter.js
@@ -4,7 +4,7 @@
 
 var _ = require('underscore');
 var Base = require('extendable-base');
-var JuttleParser = require('../parser/juttle-parser');
+var parser = require('../parser');
 var Juttle = require('../runtime/procs/procs');
 
 // List of documented, non-deprecated procs with their options.
@@ -112,7 +112,7 @@ var Autocompleter = Base.extend({
         var regions = [];
 
         try {
-            JuttleParser.parseSync(juttle, {
+            parser.parseSync(juttle, {
                 autocompleteCallback: function(region) {
                     regions.push(region);
                 }
